move inference config validation to workspace controller
