+++ api/v1beta1/workspace_validation.go
@@ -472,30 +472,6 @@ func (i *InferenceSpec) validateCreate(ctx context.Context, namespace string, in
 		errs = errs.Also(validateDuplicateName(i.Adapters, nameMap))
 	}
 
-	// check if required fields are set
-	// this check only applies to vllm runtime
-	if runtime == model.RuntimeNameVLLM {
-		func() {
-			var (
-				cmName = i.Config
-				cmNS   = namespace
-				err    error
-			)
-			if cmName == "" {
-				klog.Infof("Inference config not specified. Using default: %q", DefaultInferenceConfigTemplate)
-				cmName = DefaultInferenceConfigTemplate
-				cmNS, err = utils.GetReleaseNamespace()
-				if err != nil {
-					errs = errs.Also(apis.ErrGeneric(fmt.Sprintf("Failed to determine release namespace: %v", err), "namespace"))
-					return
-				}
-			}
-			if err := i.validateConfigMap(ctx, cmNS, cmName, instanceType); err != nil {
-				errs = errs.Also(apis.ErrGeneric(fmt.Sprintf("Failed to evaluate validateConfigMap: %v", err), "Config"))
-			}
-		}()
-	}
-
 	return errs
 }
