+++ api/v1beta1/workspace_validation_test.go
@@ -550,33 +550,6 @@ func TestInferenceSpecValidateCreate(t *testing.T) {
 	// Set environment variables
 	t.Setenv("CLOUD_PROVIDER", consts.AzureCloudName)
 	t.Setenv(consts.DefaultReleaseNamespaceEnvVar, DefaultReleaseNamespace)
-
-	// Create fake client with default ConfigMap
-	scheme := runtime.NewScheme()
-	_ = v1.AddToScheme(scheme)
-	client := fake.NewClientBuilder().WithScheme(scheme).WithRuntimeObjects(
-		defaultInferenceConfigMapManifest(),
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "valid-config",
-				Namespace: DefaultReleaseNamespace,
-			},
-			Data: map[string]string{
-				"inference_config.yaml": "a: b",
-			},
-		},
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "missing-key-config",
-				Namespace: DefaultReleaseNamespace,
-			},
-			Data: map[string]string{
-				"other_key": "some value",
-			},
-		},
-	).Build()
-	k8sclient.SetGlobalClient(client)
-
 	tests := []struct {
 		name          string
 		inferenceSpec *InferenceSpec
@@ -750,57 +723,6 @@ func TestInferenceSpecValidateCreate(t *testing.T) {
 			},
 			errContent: "Duplicate adapter source name found:",
 			expectErrs: false,
-		},
-		{
-			name: "Config specified but ConfigMap not found",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "nonexistent-config",
-			},
-			errContent: "ConfigMap 'nonexistent-config' specified in 'config' not found in namespace",
-			expectErrs: true,
-		},
-		{
-			name: "Config specified with valid ConfigMap",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "valid-config",
-			},
-			errContent: "",
-			expectErrs: false,
-		},
-		{
-			name: "ConfigMap missing required inference_config.yaml",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "missing-key-config",
-			},
-			errContent: "missing field(s): inference_config.yaml in ConfigMap",
-			expectErrs: true,
-		},
-		{
-			name: "ConfigMap missing required inference_config.yaml/transformers",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "missing-key-config",
-			},
-			runtimeName: model.RuntimeNameHuggingfaceTransformers,
 		},
 		{
 			name: "download model at runtime with access secret",
@@ -1781,206 +1703,6 @@ func TestDataDestinationValidateUpdate(t *testing.T) {
 					if !strings.Contains(errs.Error(), field) {
 						t.Errorf("validateUpdate() expected errors to contain field %s, but got %s", field, errs.Error())
 					}
-				}
-			}
-		})
-	}
-}
-
-func TestInferenceConfigMapValidation(t *testing.T) {
-	RegisterValidationTestModels()
-	ctx := context.Background()
-
-	// Create fake client with test ConfigMaps
-	scheme := runtime.NewScheme()
-	_ = v1.AddToScheme(scheme)
-	client := fake.NewClientBuilder().WithScheme(scheme).WithRuntimeObjects(
-		// ConfigMap with max-model-len set
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "valid-config-with-max-model-len",
-				Namespace: DefaultReleaseNamespace,
-			},
-			Data: map[string]string{
-				"inference_config.yaml": `
-vllm:
-  max-model-len: 2048
-  gpu-memory-utilization: 0.95
-`,
-			},
-		},
-		// ConfigMap without max-model-len set
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "invalid-config-without-max-model-len",
-				Namespace: DefaultReleaseNamespace,
-			},
-			Data: map[string]string{
-				"inference_config.yaml": `
-vllm:
-  gpu-memory-utilization: 0.95
-`,
-			},
-		},
-		// ConfigMap with empty vllm section
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "invalid-config-empty-vllm",
-				Namespace: DefaultReleaseNamespace,
-			},
-			Data: map[string]string{
-				"inference_config.yaml": `
-vllm: {}
-`,
-			},
-		},
-		// ConfigMap with vllm section missing
-		&v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "invalid-config-no-vllm",
-				Namespace: DefaultReleaseNamespace,
-			},
-			Data: map[string]string{
-				"inference_config.yaml": `
-other_field: value
-`,
-			},
-		},
-	).Build()
-	k8sclient.SetGlobalClient(client)
-
-	tests := []struct {
-		name          string
-		inferenceSpec *InferenceSpec
-		resourceSpec  *ResourceSpec
-		runtimeName   model.RuntimeName
-		errContent    string // Content expected error to include, if any
-		expectErrs    bool
-	}{
-		{
-			name: "Multi-GPU with <20GB per GPU and max-model-len set",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "valid-config-with-max-model-len",
-			},
-			resourceSpec: &ResourceSpec{
-				InstanceType: "Standard_NV12", // 2 GPUs with 8GB each (16GB total)
-				Count:        pointerToInt(1),
-			},
-			errContent: "",
-			expectErrs: false,
-		},
-		{
-			name: "Multi-GPU with <20GB per GPU and max-model-len missing",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "invalid-config-without-max-model-len",
-			},
-			resourceSpec: &ResourceSpec{
-				InstanceType: "Standard_NV12", // 2 GPUs with 8GB each (16GB total)
-				Count:        pointerToInt(1),
-			},
-			errContent: "max-model-len is required in the vllm section of inference_config.yaml when using multi-GPU instances with <20GB of memory per GPU",
-			expectErrs: true,
-		},
-		{
-			name: "Multi-GPU with <20GB per GPU and empty vllm section",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "invalid-config-empty-vllm",
-			},
-			resourceSpec: &ResourceSpec{
-				InstanceType: "Standard_NV12", // 2 GPUs with 8GB each (16GB total)
-				Count:        pointerToInt(1),
-			},
-			errContent: "max-model-len is required in the vllm section of inference_config.yaml when using multi-GPU instances with <20GB of memory per GPU",
-			expectErrs: true,
-		},
-		{
-			name: "Multi-GPU with <20GB per GPU and vllm section missing",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "invalid-config-no-vllm",
-			},
-			resourceSpec: &ResourceSpec{
-				InstanceType: "Standard_NV12", // 2 GPUs with 8GB each (16GB total)
-				Count:        pointerToInt(1),
-			},
-			errContent: "max-model-len is required in the vllm section of inference_config.yaml when using multi-GPU instances with <20GB of memory per GPU",
-			expectErrs: true,
-		},
-		{
-			name: "Single-GPU instance (no max-model-len required)",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "invalid-config-without-max-model-len",
-			},
-			resourceSpec: &ResourceSpec{
-				InstanceType: "Standard_NV6", // 1 GPU with 8GB
-				Count:        pointerToInt(1),
-			},
-			errContent: "",
-			expectErrs: false,
-		},
-		{
-			name: "Multi-GPU with >=20GB per GPU (no max-model-len required)",
-			inferenceSpec: &InferenceSpec{
-				Preset: &PresetSpec{
-					PresetMeta: PresetMeta{
-						Name: ModelName("test-validation"),
-					},
-				},
-				Config: "invalid-config-without-max-model-len",
-			},
-			resourceSpec: &ResourceSpec{
-				InstanceType: "Standard_NC48ads_A100_v4", // 2 GPUs with 80GB each (160GB total)
-				Count:        pointerToInt(1),
-			},
-			errContent: "",
-			expectErrs: false,
-		},
-	}
-
-	t.Setenv("CLOUD_PROVIDER", consts.AzureCloudName)
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			// Validate the inference spec
-			runtimeName := model.RuntimeNameVLLM
-			if tc.runtimeName != "" {
-				runtimeName = tc.runtimeName
-			}
-			errs := tc.inferenceSpec.validateCreate(ctx, DefaultReleaseNamespace, tc.resourceSpec.InstanceType, runtimeName)
-			hasErrs := errs != nil
-
-			if hasErrs != tc.expectErrs {
-				t.Errorf("validateCreate() errors = %v, expectErrs %v", errs, tc.expectErrs)
-			}
-
-			if hasErrs && tc.errContent != "" {
-				errMsg := errs.Error()
-				if !strings.Contains(errMsg, tc.errContent) {
-					t.Errorf("validateCreate() error message = %v, expected to contain = %v", errMsg, tc.errContent)
 				}
 			}
 		})
